server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
pyjwt
requests
